fix SMPS preparser (mostly) to enumerates DAC volume levels properly
This involves following SMPS loops, so that all played notes are processed
with their correct volume.
It will still break if subroutines are used.

Following loops has a risk of freezing (due to improperly nested loops),
so the special code is enable ONLY on drum tracks with VGM logging enabled.