read/use register array of instrument libraries
This adds support for mixed instrument modes, i.e. the instrument format
used by a song's private instruments and FM drums can be different from
the format of the global instrument library now.

This change was required to support Darwin 4081. It uses the interleaved
instrument format for FM drums and the "custom" setting for the instrument
library.